src/requirements.txt
aiofiles==0.8.0
aiosqlite==0.17.0
asgiref==3.4.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chance==0.110
charset-normalizer==2.0.9
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.1
django-allauth==0.22.0
django-auto-serializer==0.4.2
django-countries==7.2.1
django-crispy-forms==1.13.0
django-jalali==5.0.0
django-jalali-date==1.0.1
django-model-serializer==0.1.0
django-multi-form-view==2.0.1
django-rest-multiple-models==2.1.3
django-serializer==0.0.8
django-serializer-json==0.10
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-serializer-extensions==2.0.1
djangorestframework-serializer-mixins==0.1.0
djangorestframework-serializer-query-optimizations==0.1.0
greenlet==1.1.2
hashids==1.3.1
httptools==0.3.0
hupper==1.10.3
idna==3.3
inflection==0.5.1
iso8601==0.1.16
jdatetime==3.7.0
jsonschema==4.3.3
multidict==5.2.0
oauthlib==3.1.1
PasteDeploy==2.1.1
Pillow==9.0.0
plaster==1.0
plaster-pastedeploy==0.7
psycopg2==2.9.3
pycparser==2.21
pydot==1.4.2
PyJWT==2.3.0
pyparsing==3.0.6
pypika-tortoise==0.1.2
pyramid==2.0
pyramid-mailer==0.15.1
pyramid-rest-framework==0.0.6
pyrsistent==0.18.0
python3-openid==3.2.0
pytz==2021.3
repoze.sendmail==4.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rest-framework-custom-viewsets==0.0.4
rest-framework-dyn-serializer==1.3.1
rest-framework-ember==1.3.2
rfc3987==1.3.8
sanic==21.12.0
sanic-rest-framework==1.2.1
sanic-routing==0.7.2
Serializer==0.2.1
signals==0.0.2
simplejson==3.17.6
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
SQLAHelper==1.0
SQLAlchemy==1.4.29
sqlalchemy-schemadisplay==1.3
sqlparse==0.4.2
strict-rfc3339==0.7
tortoise-orm==0.18.0
transaction==3.0.1
translationstring==1.4
typing_extensions==4.0.1
tzdata==2021.5
ujson==5.1.0
urllib3==1.26.7
venusian==3.0.0
WebOb==1.8.7
websockets==10.1
whitenoise==5.3.0
zope.deprecation==4.4.0
zope.interface==5.4.0
zope.sqlalchemy==1.6
